refactor: 컴포넌트와 스토리 파일 colocation 구조로 전환
