+++ my-app/src/components/Card/Card.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import Card from "../components/Card";
+import Card from "./Card";
 
 const meta = {
   title: "Components/Card",
