+++ my-app/src/components/Card/masterLogo.png
