+++ my-app/src/components/CardNumberInputSection/CardNumberInputSection.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { expect, fn, userEvent, within } from "storybook/test";
-import CardNumberInputSection from "../components/CardNumberInputSection";
+import CardNumberInputSection from "./CardNumberInputSection";
 
 const meta = {
   title: "Components/CardNumberInputSection",
