+++ my-app/src/components/CardNumberInputSection/CardNumberInputSection.tsx
@@ -1,14 +1,15 @@
+import InputSectionLayout from "../InputSectionLayout/InputSectionLayout";
+import ValidatedInputGroup from "../ValidatedInputGroup/ValidatedInputGroup";
 import { useState } from "react";
-import InputSectionLayout from "./InputSectionLayout";
-import ValidatedInputGroup from "./ValidatedInputGroup";
+import { decideBrandName } from "../../utils/decideBrandName";
 
-const ExpiryDateInputSection = ({
+const CardNumberInputSection = ({
   onValueHandler,
 }: {
   onValueHandler: (cardInfo: string[], brand?: string) => void;
 }) => {
   const [inputValues, setInputValues] = useState<string[]>([]);
-  const [errorMessage, setErrorMessage] = useState("");
+  const [errorMessage, setErrorMessage] = useState<string>("");
   const [errorIndex, setErrorIndex] = useState<number>(-1);
 
   const onChange = (index: number, value: string) => {
@@ -32,9 +33,9 @@ const ExpiryDateInputSection = ({
         message = "숫자만 입력 가능합니다";
         break;
       }
-      if (i === 0 && !/^(0[1-9]|1[0-2])$/.test(value)) {
+      if (i === 0 && decideBrandName(value) === "") {
         errorIndex = i;
-        message = "유효한 날짜를 입력해주세요";
+        message = "이 카드 브랜드는 지원하지 않습니다.";
         break;
       }
     }
@@ -45,9 +46,9 @@ const ExpiryDateInputSection = ({
 
   return (
     <InputSectionLayout
-      title="카드 유효기간을 입력해주세요"
-      message="월/년도(MMYY)를 순서대로 입력해 주세요."
-      tag="유효기간"
+      title="결제할 카드 번호를 입력해 주세요"
+      message="본인 명의의 카드만 결제 가능합니다."
+      tag="카드 번호"
     >
       <ValidatedInputGroup
         onChange={onChange}
@@ -55,10 +56,10 @@ const ExpiryDateInputSection = ({
         errorMessage={errorMessage}
         errorIndex={errorIndex}
         values={inputValues}
-        inputOption={{ count: 2, maxLength: 2, placeHolder: ["MM", "YY"] }}
+        inputOption={{ count: 4, maxLength: 4, placeHolder: ["1234", "1234", "1234", "1234"] }}
       />
     </InputSectionLayout>
   );
 };
 
-export default ExpiryDateInputSection;
+export default CardNumberInputSection;
