+++ my-app/src/components/ExpiryDateInputSection/ExpiryDateInputSection.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { expect, fn, userEvent, within } from "storybook/test";
-import ExpiryDateInputSection from "../components/ExpiryDateInputSection";
+import ExpiryDateInputSection from "./ExpiryDateInputSection";
 
 const meta = {
   title: "Components/ExpiryDateInputSection",
