+++ my-app/src/components/ExpiryDateInputSection/ExpiryDateInputSection.tsx
@@ -1,15 +1,14 @@
-import InputSectionLayout from "./InputSectionLayout";
-import ValidatedInputGroup from "./ValidatedInputGroup";
 import { useState } from "react";
-import { decideBrandName } from "../utils/decideBrandName";
+import InputSectionLayout from "../InputSectionLayout/InputSectionLayout";
+import ValidatedInputGroup from "../ValidatedInputGroup/ValidatedInputGroup";
 
-const CardNumberInputSection = ({
+const ExpiryDateInputSection = ({
   onValueHandler,
 }: {
   onValueHandler: (cardInfo: string[], brand?: string) => void;
 }) => {
   const [inputValues, setInputValues] = useState<string[]>([]);
-  const [errorMessage, setErrorMessage] = useState<string>("");
+  const [errorMessage, setErrorMessage] = useState("");
   const [errorIndex, setErrorIndex] = useState<number>(-1);
 
   const onChange = (index: number, value: string) => {
@@ -33,9 +32,9 @@ const CardNumberInputSection = ({
         message = "숫자만 입력 가능합니다";
         break;
       }
-      if (i === 0 && decideBrandName(value) === "") {
+      if (i === 0 && !/^(0[1-9]|1[0-2])$/.test(value)) {
         errorIndex = i;
-        message = "이 카드 브랜드는 지원하지 않습니다.";
+        message = "유효한 날짜를 입력해주세요";
         break;
       }
     }
@@ -46,9 +45,9 @@ const CardNumberInputSection = ({
 
   return (
     <InputSectionLayout
-      title="결제할 카드 번호를 입력해 주세요"
-      message="본인 명의의 카드만 결제 가능합니다."
-      tag="카드 번호"
+      title="카드 유효기간을 입력해주세요"
+      message="월/년도(MMYY)를 순서대로 입력해 주세요."
+      tag="유효기간"
     >
       <ValidatedInputGroup
         onChange={onChange}
@@ -56,10 +55,10 @@ const CardNumberInputSection = ({
         errorMessage={errorMessage}
         errorIndex={errorIndex}
         values={inputValues}
-        inputOption={{ count: 4, maxLength: 4, placeHolder: ["1234", "1234", "1234", "1234"] }}
+        inputOption={{ count: 2, maxLength: 2, placeHolder: ["MM", "YY"] }}
       />
     </InputSectionLayout>
   );
 };
 
-export default CardNumberInputSection;
+export default ExpiryDateInputSection;
