+++ my-app/src/components/InputSectionLayout/InputSectionLayout.tsx
